requirements.txt
polars>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
cryptography>=42.0.0
python-dotenv>=1.0.0

